lab3/requirements.txt
# run pip install -r requirements.txt to automatically install necessary packages
des
Pillow